+++ NEWS.md
@@ -1,3 +1,9 @@
+* Added `keyword` parameter to `dataset()` function, enabling simple keyword search of dataset metadata via the `/dataset/search` endpoint.
+
+# robis 2.11.3
+
+Warnings and notes fixes.
+
 # robis 2.11.3
 * Added `keyword` parameter to `dataset()` function, enabling simple keyword search of dataset metadata via the `/dataset/search2` endpoint.